+++ Ryujinx.HLE/HOS/Services/Vi/RootService/ApplicationDisplayService/ISystemDisplayService.cs
@@ -47,8 +47,8 @@ namespace Ryujinx.HLE.HOS.Services.Vi.RootService.ApplicationDisplayService
             (ulong width, ulong height) = AndroidSurfaceComposerClient.GetDisplayInfo(context, displayId);
 
             // TODO: De-hardcode resolution.
-            context.ResponseData.Write(width);
-            context.ResponseData.Write(height);
+            context.ResponseData.Write((uint)width);
+            context.ResponseData.Write((uint)height);
             context.ResponseData.Write(60.0f);
             context.ResponseData.Write(0);
